requirements.txt
fastapi
uvicorn
gunicorn 
sqlalchemy
sqlmodel
pyjwt
pydantic
python-decouple
psycopg2-binary
bcrypt
python-multipart 